add `x-enumNames` data to generated OpenAPI spec
OpenAPI doesn't have a single, straightforward way to give enums descriptors. [This stackoverflow answer](https://stackoverflow.com/a/66471626) goes through some of the options

This project will be using `ng-openapi-gen`, so we need to be sure our enum strategy is compatible with that. `ng-openapi-gen` supports the `x-enumNames` extension, so that's what we add here

While the `x-enumNames` aren't shown in the swagger UI, they can be viewed in the generated `swagger.json` file linked in the header of the swagger UI page